+++ ark-common/src/main/java/au/org/theark/core/model/lims/entity/BiodataGroupCriteria.java
@@ -2,7 +2,6 @@ package au.org.theark.core.model.lims.entity;
 
 // Generated 15/06/2011 1:22:58 PM by Hibernate Tools 3.3.0.GA
 
-import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
@@ -23,7 +22,7 @@ import au.org.theark.core.model.Constants;
 public class BiodataGroupCriteria implements java.io.Serializable
 {
 
-	private int					id;
+	private Long					id;
 	private BiodataGroup		biodataGroup;
 	private BiodataCriteria	biodataCriteria;
 
@@ -31,7 +30,7 @@ public class BiodataGroupCriteria implements java.io.Serializable
 	{
 	}
 
-	public BiodataGroupCriteria(int id, BiodataGroup biodataGroup, BiodataCriteria biodataCriteria)
+	public BiodataGroupCriteria(Long id, BiodataGroup biodataGroup, BiodataCriteria biodataCriteria)
 	{
 		this.id = id;
 		this.biodataGroup = biodataGroup;
@@ -39,14 +38,14 @@ public class BiodataGroupCriteria implements java.io.Serializable
 	}
 
 	@Id
-	@SequenceGenerator(name="biodatagroupcriteria_generator", sequenceName="BIODATAGROUPCRITERIA_SEQUENCE")
-	@GeneratedValue(strategy=GenerationType.AUTO, generator = "biodatagroupcriteria_generator")
-	public int getId()
+	@SequenceGenerator(name = "biodatagroupcriteria_generator", sequenceName = "BIODATAGROUPCRITERIA_SEQUENCE")
+	@GeneratedValue(strategy = GenerationType.AUTO, generator = "biodatagroupcriteria_generator")
+	public Long getId()
 	{
 		return this.id;
 	}
 
-	public void setId(int id)
+	public void setId(Long id)
 	{
 		this.id = id;
 	}
